refactor: extract reader, caching, and sql partials
